Fix off-by-ones and update shear buckling calculation method
Fix off-by-ones when it comes to the length of sfds and bridges to make
things more consistent. Now the sfd for a bridge of length L has exactly
L points instead of L + 1.

Use a variable distance between diaphragms for the shear buckling
calculation method instead of a single fixed one.